feat(global): :sparkles: app builder modelManager compnents handled by single component
app builder modelManager compnents handled by single component

Ref: #45
